requirements.txt
absl-py==2.3.1
addict==2.4.0
aiodns==3.1.1
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aioitertools==0.11.0
aiosignal==1.3.2
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
anytree==2.13.0
argcomplete==3.6.3
arm_pytorch_utilities==0.4.3
asteval==1.0.6
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.1.0
autocommand==2.2.2
azure-core==1.28.0
azure-identity==1.13.0
azure-storage-blob==12.17.0
babel==2.17.0
backports.tarfile==1.2.0
beautifulsoup4==4.14.3
blis==1.3.3
boto3==1.36.1
botocore==1.36.1
cachetools==6.2.4
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.3.2
cheroot==11.1.2
CherryPy==18.10.0
click==8.1.7
cloudpathlib==0.23.0
cloudpickle==3.1.1
cmeel==0.57.3
cmeel-assimp==5.4.3.1
cmeel-boost==1.83.0
cmeel-console-bridge==1.0.2.3
cmeel-octomap==1.10.0
cmeel-qhull==8.0.2.1
cmeel-tinyxml==2.6.2.3
cmeel-urdfdom==3.1.1.1
cmeel-zlib==1.3.1
coloredlogs==15.0.1
colorlog==6.10.1
comm==0.2.3
confection==0.1.5
construct==2.10.68
contourpy==1.3.1
coverage==7.4.4
cryptography==44.0.0
csvw==3.7.0
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cycler==0.11.0
cymem==2.0.13
Cython==3.2.3
cython_bbox==0.1.5
decorator==4.4.2
dex_retargeting==0.4.6
distro==1.9.0
dlinfo==2.0.0
docopt==0.6.2
easydict==1.13
egl_probe==1.0.2
eigenpy==3.5.1
einops==0.8.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
espeakng-loader==0.2.4
executing==2.2.0
Farama-Notifications==0.0.4
fastapi==0.115.7
ffmpeg-python==0.2.0
filelock==3.13.1
flaky==3.8.1
flatbuffers==25.12.19
flatdict==4.0.1
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.6.1
ftfy==6.3.1
future==1.0.0
fvcore==0.1.5.post20221221
# Editable install with no version control (g1_robot_sim==0.0.1)
-e /home/robot/thang_project/g1_sim
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.45.0
google-genai==1.56.0
grpcio==1.74.0
gunicorn==23.0.0
gym==0.23.1
gym-notices==0.1.0
gymnasium==1.2.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.9
hidapi==0.14.0.post2
hpp-fcl==2.4.4
httpcore==1.0.9
httptools==0.6.1
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
idna-ssl==1.1.0
imageio==2.34.1
imageio-ffmpeg==0.4.5
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_importlib-metadata_1747934053/work
iniconfig==2.1.0
insightface==0.7.3
iopath==0.1.10
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.5
-e git+https://github.com/isaac-sim/IsaacLab.git@1b2d6eaf0c887380387a4bd3fe154e737aacadcc#egg=isaaclab&subdirectory=source/isaaclab
-e git+https://github.com/isaac-sim/IsaacLab.git@1b2d6eaf0c887380387a4bd3fe154e737aacadcc#egg=isaaclab_assets&subdirectory=source/isaaclab_assets
-e git+https://github.com/isaac-sim/IsaacLab.git@1b2d6eaf0c887380387a4bd3fe154e737aacadcc#egg=isaaclab_mimic&subdirectory=source/isaaclab_mimic
-e git+https://github.com/isaac-sim/IsaacLab.git@1b2d6eaf0c887380387a4bd3fe154e737aacadcc#egg=isaaclab_rl&subdirectory=source/isaaclab_rl
-e git+https://github.com/isaac-sim/IsaacLab.git@1b2d6eaf0c887380387a4bd3fe154e737aacadcc#egg=isaaclab_tasks&subdirectory=source/isaaclab_tasks
isodate==0.6.1
jaraco.collections==5.2.1
jaraco.context==6.0.2
jaraco.functools==4.4.0
jaraco.text==4.0.0
jax==0.8.2
jaxlib==0.8.2
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
junitparser==4.0.2
jupyterlab_widgets==3.0.15
kiwisolver==1.4.4
kokoro==0.9.4
language-tags==1.2.0
lapx==0.9.3
lazy_loader==0.4
llvmlite==0.42.0
loguru==0.7.3
loop-rate-limiters==1.2.0
lxml==5.4.0
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.11
misaki==0.9.4
ml_dtypes==0.5.3
more-itertools==10.8.0
moviepy==1.0.3
mpmath==1.3.0
msal==1.23.0
msal-extensions==1.0.0
msgpack==1.1.1
multidict==6.1.0
murmurhash==1.0.15
nest-asyncio==1.5.6
networkx==3.3
nlopt==2.7.1
num2words==0.5.14
numba==0.59.1
numpy==1.26.4
numpy-quaternion==2023.0.3
nvidia-cublas-cu12==12.8.3.14
nvidia-cuda-cupti-cu12==12.8.57
nvidia-cuda-nvrtc-cu12==12.8.61
nvidia-cuda-runtime-cu12==12.8.57
nvidia-cudnn-cu12==9.7.1.26
nvidia-cufft-cu12==11.3.3.41
nvidia-cufile-cu12==1.13.0.11
nvidia-curand-cu12==10.3.9.55
nvidia-cusolver-cu12==11.7.2.55
nvidia-cusparse-cu12==12.5.7.53
nvidia-cusparselt-cu12==0.6.3
nvidia-lula-no-cuda==0.10.1
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.8.55
nvidia-srl-base==1.3.0
nvidia-srl-math==1.0.0
nvidia-srl-usd==1.0.0
nvidia-srl-usd-to-urdf==1.0.1
nvsmi==0.4.2
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.20.0
onnxruntime==1.23.2
onnxruntime-gpu==1.23.2
openai-clip==1.0.1
opencv-contrib-python==4.12.0.88
opencv-python==4.6.0.66
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
osqp==0.6.5
packaging==23.0
pandas==2.1.4
parso==0.8.5
pexpect==4.9.0
phonemizer-fork==3.3.2
pillow==11.0.0
pin==2.7.0
pin-pink==3.1.0
Pint==0.20.1
platformdirs==4.4.0
plotly==5.3.1
pluggy==1.6.0
polars==1.36.1
polars-runtime-32==1.36.1
portalocker==2.7.0
portend==3.2.1
preshed==3.0.12
prettytable==3.3.0
proglog==0.1.12
prompt_toolkit==3.0.52
propcache==0.2.1
protobuf==3.20.3
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycares==4.8.0
pycolmap==3.13.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pyglet==1.5.31
Pygments==2.19.2
pyparsing==3.0.9
pyperclip==1.8.0
pypng==0.20220715.0
PySocks==1.7.1
pytest==8.4.1
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytorch-seed==0.2.0
pytorch_mppi==0.8.0
pytransform3d==3.14.2
pytz==2024.1
PyYAML==6.0.2
qdldl==0.1.7.post1
qpsolvers==4.8.1
qrcode==7.4.2
quadprog==0.1.13
ray==2.49.0
rdflib==7.5.0
referencing==0.36.2
regex==2025.7.34
requests==2.32.3
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==14.1.0
rl_games @ git+https://github.com/isaac-sim/rl_games.git@0b15191ca28e19be1854297c38b5a392749abac6
robomimic @ git+https://github.com/ARISE-Initiative/robomimic.git@7c66e7a41b5d9dcc905b1a68346bfee1b49b79c9
rpds-py==0.27.1
rsa==4.9.1
rsl-rl-lib==2.3.3
Rtree==1.3.0
s3transfer==0.6.1
safetensors==0.6.2
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.14.1
seaborn==0.13.2
segments==2.3.0
selenium==4.14.0
sentry-sdk==2.29.1
setproctitle==1.3.6
setuptools-scm==8.2.0
simsimd==6.5.12
six==1.17.0
skrl==1.4.3
smart_open==7.5.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.3
soundfile==0.13.1
soupsieve==2.8.1
spacy==3.8.11
spacy-curated-transformers==0.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
stable_baselines3==2.7.0
stack-data==0.6.3
starlette==0.45.3
stringzilla==4.6.0
sympy==1.13.3
tabulate==0.9.0
tempora==5.8.1
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
termcolor==2.5.0
thinc==8.3.10
threadpoolctl==3.6.0
tifffile==2025.12.20
tokenizers==0.21.4
toml==0.10.2
tomli==2.2.1
torch==2.7.0+cu128
torchaudio==2.7.0+cu128
torchreid==0.2.5
torchvision==0.22.0+cu128
tornado==6.5
tqdm==4.67.1
traitlets==5.14.3
transformers==4.55.4
transforms3d==0.4.2
trimesh==4.10.1
triton==3.3.0
typer-slim==0.21.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
ultralytics==8.3.241
ultralytics-thop==2.0.18
-e git+https://github.com/thinhhd-cyber/unitree_isaac.git@6f139a910d40f834d04fa277491d3fe7b7b42bdd#egg=unitree_rl_lab&subdirectory=source/unitree_rl_lab
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.29.0
wandb==0.21.1
warp-lang==1.8.1
wasabi==1.1.3
watchdog==4.0.0
wcwidth==0.2.13
weasel==0.4.3
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wis3d==1.0.1
wrapt==1.16.0
yacs==0.1.8
yarl==1.18.3
zc.lockfile==4.0
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1749421620841/work
